Splitted composition from StockToolsBindings.
