+++ ts/Stock/StockToolsBindings.ts
@@ -29,17 +29,13 @@ import type NavigationBindingsOptions from '../Extensions/Annotations/Navigation
 import type PointerEvent from '../Core/PointerEvent';
 import type { SeriesTypeOptions } from '../Core/Series/SeriesType';
 
-import D from '../Core/DefaultOptions.js';
-const { setOptions } = D;
 import H from '../Core/Globals.js';
 import NavigationBindings from '../Extensions/Annotations/NavigationBindings.js';
 import NBU from '../Extensions/Annotations/NavigationBindingsUtilities.js';
 const { getAssignedAxis } = NBU;
 import { Palette } from '../Core/Color/Palettes.js';
-import Series from '../Core/Series/Series.js';
 import STU from './StockToolsUtilities.js';
 const {
-    indicatorsWithAxes,
     addFlagFromForm,
     attractToPoint,
     isNotNavigatorYAxis,
@@ -52,353 +48,38 @@ const {
 import U from '../Core/Utilities.js';
 import FibonacciTimeZones from '../Extensions/Annotations/Types/FibonacciTimeZones';
 const {
-    correctFloat,
-    defined,
-    extend,
     fireEvent,
-    isNumber,
-    merge,
-    pick
+    merge
 } = U;
 
-/**
- * Internal types
- * @private
- */
-declare global {
-    namespace Highcharts {
-        interface StockToolsNavigationBindings extends NavigationBindings {
-            toggledAnnotations?: boolean;
-            utils: typeof STU;
-            verticalCounter?: number;
-            /** @requires modules/stock-tools */
-            getYAxisPositions(
-                yAxes: Array<AxisType>,
-                plotHeight: number,
-                defaultHeight: number,
-                removedYAxisProps?: AxisPositions
-            ): YAxisPositions;
-            /** @requires modules/stock-tools */
-            getYAxisResizers(
-                yAxes: Array<AxisType>
-            ): Array<NavigationBindingsResizerObject>;
-            /** @requires modules/stock-tools */
-            recalculateYAxisPositions(
-                positions: Array<Record<string, number>>,
-                changedSpace: number,
-                modifyHeight?: boolean,
-                adder?: number
-            ): Array<Record<string, number>>;
-            /** @requires modules/stock-tools */
-            resizeYAxes(
-                removedYAxisProps?: Highcharts.AxisPositions
-            ): void;
-        }
-        interface NavigationBindingsAttractionObject {
-            x: number;
-            y: number;
-            below: boolean;
-            series: Series;
-            xAxis: number;
-            yAxis: number;
-        }
-
-        interface YAxisPositions {
-            positions: Array<Record<string, number>>;
-            allAxesHeight: number;
-        }
-        interface AxisPositions {
-            top: string;
-            height: string;
-        }
-        interface NavigationBindingsResizerObject {
-            controlledAxis?: Record<string, Array<number>>;
-            enabled: boolean;
-        }
-        interface StockToolsFieldsObject {
-            [key: string]: any;
-        }
-    }
-}
+/* *
+ *
+ *  Declarations
+ *
+ * */
 
 declare module '../Extensions/Annotations/NavigationBindingsLike' {
     interface NavigationBindingsLike {
-        utils: Partial<typeof STU>;
+        toggledAnnotations?: boolean;
+        verticalCounter?: number;
     }
 }
 
-// Extends NavigationBindigs to support indicators and resizers:
-extend<NavigationBindings|Highcharts.StockToolsNavigationBindings>(NavigationBindings.prototype, {
-    /* eslint-disable valid-jsdoc */
-    /**
-     * Get current positions for all yAxes. If new axis does not have position,
-     * returned is default height and last available top place.
-     *
-     * @private
-     * @function Highcharts.NavigationBindings#getYAxisPositions
-     *
-     * @param {Array<Highcharts.Axis>} yAxes
-     *        Array of yAxes available in the chart.
-     *
-     * @param {number} plotHeight
-     *        Available height in the chart.
-     *
-     * @param {number} defaultHeight
-     *        Default height in percents.
-     *
-     * @param {Highcharts.AxisPositions} removedYAxisProps
-     *        Height and top value of the removed yAxis in percents.
-     *
-     * @return {Highcharts.YAxisPositions}
-     *         An object containing an array of calculated positions
-     *         in percentages. Format: `{top: Number, height: Number}`
-     *         and maximum value of top + height of axes.
-     */
-    getYAxisPositions: function (
-        yAxes: Array<AxisType>,
-        plotHeight: number,
-        defaultHeight: number,
-        removedYAxisProps?: Highcharts.AxisPositions
-    ): Highcharts.YAxisPositions {
-        let positions: Array<Record<string, number>>|undefined,
-            allAxesHeight = 0,
-            previousAxisHeight: number,
-            removedHeight: number,
-            removedTop: number;
-        /** @private */
-        function isPercentage(prop: number | string | undefined): boolean {
-            return defined(prop) && !isNumber(prop) && (prop.match('%') as any);
-        }
-
-        if (removedYAxisProps) {
-            removedTop = correctFloat(
-                (parseFloat(removedYAxisProps.top) / 100)
-            );
-            removedHeight = correctFloat(
-                (parseFloat(removedYAxisProps.height) / 100)
-            );
-        }
-
-        positions = yAxes.map(function (yAxis: AxisType, index: number): Record<string, number> {
-            let height = correctFloat(isPercentage(yAxis.options.height) ?
-                    parseFloat(yAxis.options.height as any) / 100 :
-                    yAxis.height / plotHeight),
-                top = correctFloat(isPercentage(yAxis.options.top) ?
-                    parseFloat(yAxis.options.top as any) / 100 :
-                    (yAxis.top - yAxis.chart.plotTop) / plotHeight);
-
-            if (!removedHeight) {
-                // New axis' height is NaN so we can check if
-                // the axis is newly created this way
-                if (!isNumber(height)) {
-                    // Check if the previous axis is the
-                    // indicator axis (every indicator inherits from sma)
-                    height = yAxes[index - 1].series
-                        .every((s: Series): boolean => s.is('sma')) ?
-                        previousAxisHeight : defaultHeight / 100;
-                }
-
-                if (!isNumber(top)) {
-                    top = allAxesHeight;
-                }
-
-                previousAxisHeight = height;
-
-                allAxesHeight = correctFloat(Math.max(
-                    allAxesHeight,
-                    (top || 0) + (height || 0)
-                ));
-            } else {
-                // Move all axes which were below the removed axis up.
-                if (
-                    top > removedTop
-                ) {
-                    top -= removedHeight;
-                }
-
-                allAxesHeight = Math.max(
-                    allAxesHeight,
-                    (top || 0) + (height || 0)
-                );
-            }
-
-            return {
-                height: height * 100,
-                top: top * 100
-            };
-        });
-
-        return { positions, allAxesHeight };
-    },
-
-    /**
-     * Get current resize options for each yAxis. Note that each resize is
-     * linked to the next axis, except the last one which shouldn't affect
-     * axes in the navigator. Because indicator can be removed with it's yAxis
-     * in the middle of yAxis array, we need to bind closest yAxes back.
-     *
-     * @private
-     * @function Highcharts.NavigationBindings#getYAxisResizers
-     *
-     * @param {Array<Highcharts.Axis>} yAxes
-     *        Array of yAxes available in the chart
-     *
-     * @return {Array<object>}
-     *         An array of resizer options.
-     *         Format: `{enabled: Boolean, controlledAxis: { next: [String]}}`
-     */
-    getYAxisResizers: function (
-        yAxes: Array<AxisType>
-    ): Array<Highcharts.NavigationBindingsResizerObject> {
-        const resizers: Array<Highcharts.NavigationBindingsResizerObject> = [];
-
-        yAxes.forEach(function (_yAxis: AxisType, index: number): void {
-            const nextYAxis = yAxes[index + 1];
-
-            // We have next axis, bind them:
-            if (nextYAxis) {
-                resizers[index] = {
-                    enabled: true,
-                    controlledAxis: {
-                        next: [
-                            pick(
-                                nextYAxis.options.id,
-                                nextYAxis.options.index as any
-                            )
-                        ]
-                    }
-                };
-            } else {
-                // Remove binding:
-                resizers[index] = {
-                    enabled: false
-                };
-            }
-        });
-
-        return resizers;
-    },
-    /**
-     * Resize all yAxes (except navigator) to fit the plotting height. Method
-     * checks if new axis is added, if the new axis will fit under previous
-     * axes it is placed there. If not, current plot area is scaled
-     * to make room for new axis.
-     *
-     * If axis is removed, the current plot area streaches to fit into 100%
-     * of the plot area.
-     *
-     * @private
-     * @function Highcharts.NavigationBindings#resizeYAxes
-     * @param {Highcharts.AxisPositions} [removedYAxisProps]
-     *
-     *
-     */
-    resizeYAxes: function (
-        this: Highcharts.StockToolsNavigationBindings,
-        removedYAxisProps?: Highcharts.AxisPositions
-    ): void {
-        // The height of the new axis before rescalling. In %, but as a number.
-        const defaultHeight = 20;
-        const chart = this.chart,
-            // Only non-navigator axes
-            yAxes = chart.yAxis.filter(isNotNavigatorYAxis),
-            plotHeight = chart.plotHeight,
-            // Gather current heights (in %)
-            { positions, allAxesHeight } = this.getYAxisPositions(
-                yAxes,
-                plotHeight,
-                defaultHeight,
-                removedYAxisProps
-            ),
-            resizers = this.getYAxisResizers(yAxes);
-
-        // check if the axis is being either added or removed and
-        // if the new indicator axis will fit under existing axes.
-        // if so, there is no need to scale them.
-        if (
-            !removedYAxisProps &&
-            allAxesHeight <= correctFloat(0.8 + defaultHeight / 100)
-        ) {
-            positions[positions.length - 1] = {
-                height: defaultHeight,
-                top: correctFloat(allAxesHeight * 100 - defaultHeight)
-            };
-        } else {
-            positions.forEach(function (
-                position: Record<string, number>
-            ): void {
-                position.height = (
-                    position.height / (allAxesHeight * 100)
-                ) * 100;
-                position.top = (position.top / (allAxesHeight * 100)) * 100;
-            });
-        }
-
-        positions.forEach(function (
-            position: Record<string, number>, index: number
-        ): void {
-            yAxes[index].update({
-                height: position.height + '%',
-                top: position.top + '%',
-                resize: resizers[index],
-                offset: 0
-            }, false);
-        });
-    },
-
-    /**
-     * Utility to modify calculated positions according to the remaining/needed
-     * space. Later, these positions are used in `yAxis.update({ top, height })`
-     *
-     * @private
-     * @function Highcharts.NavigationBindings#recalculateYAxisPositions
-     * @param {Array<Highcharts.Dictionary<number>>} positions
-     * Default positions of all yAxes.
-     * @param {number} changedSpace
-     * How much space should be added or removed.
-     * @param {boolean} modifyHeight
-     * Update only `top` or both `top` and `height`.
-     * @param {number} adder
-     * `-1` or `1`, to determine whether we should add or remove space.
-     *
-     * @return {Array<object>}
-     *         Modified positions,
-     */
-    recalculateYAxisPositions: function (
-        positions: Array<Record<string, number>>,
-        changedSpace: number,
-        modifyHeight?: boolean,
-        adder?: number
-    ): Array<Record<string, number>> {
-        positions.forEach(function (
-            position: Record<string, number>,
-            index: number
-        ): void {
-            const prevPosition = positions[index - 1];
-
-            position.top = !prevPosition ? 0 :
-                correctFloat(prevPosition.height + prevPosition.top);
-
-            if (modifyHeight) {
-                position.height = correctFloat(
-                    position.height + (adder as any) * changedSpace
-                );
-            }
-        });
-
-        return positions;
-    }
-    /* eslint-enable valid-jsdoc */
-});
+/* *
+ *
+ *  Constants
+ *
+ * */
 
 /**
+ * @sample {highstock} stock/stocktools/custom-stock-tools-bindings
+ *         Custom stock tools bindings
+ *
  * @type         {Highcharts.Dictionary<Highcharts.NavigationBindingsOptionsObject>}
  * @since        7.0.0
  * @optionparent navigation.bindings
- *   @sample {highstock} stock/stocktools/custom-stock-tools-bindings
- *     Custom stock tools bindings
  */
-const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
+const StockToolsBindings: Record<string, NavigationBindingsOptions> = {
     // Line type annotations:
     /**
      * A segment annotation bindings. Includes `start` and one event in `steps`
@@ -1490,7 +1171,7 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
         // eslint-disable-next-line valid-jsdoc
         /** @ignore-option */
         start: function (
-            this: Highcharts.StockToolsNavigationBindings,
+            this: NavigationBindings,
             e: PointerEvent
         ): void {
             let closestPoint = attractToPoint(e, this.chart),
@@ -2193,7 +1874,7 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
         className: 'highcharts-indicators',
         // eslint-disable-next-line valid-jsdoc
         /** @ignore-option */
-        init: function (this: Highcharts.StockToolsNavigationBindings): void {
+        init: function (this: NavigationBindings): void {
             const navigation = this;
 
             fireEvent(
@@ -2204,7 +1885,7 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
                     options: {},
                     // Callback on submit:
                     onSubmit: function (data: any): void {
-                        navigation.utils.manageIndicators.call(
+                        manageIndicators.call(
                             navigation,
                             data
                         );
@@ -2226,7 +1907,7 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
         // eslint-disable-next-line valid-jsdoc
         /** @ignore-option */
         init: function (
-            this: Highcharts.StockToolsNavigationBindings,
+            this: NavigationBindings,
             button: HTMLDOMElement
         ): void {
             const chart = this.chart,
@@ -2236,7 +1917,7 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
             this.toggledAnnotations = !this.toggledAnnotations;
 
             (chart.annotations || []).forEach(function (
-                this: Highcharts.StockToolsNavigationBindings,
+                this: NavigationBindings,
                 annotation: any
             ): void {
                 annotation.setVisibility(!this.toggledAnnotations);
@@ -2279,7 +1960,7 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
         // eslint-disable-next-line valid-jsdoc
         /** @ignore-option */
         init: function (
-            this: Highcharts.StockToolsNavigationBindings,
+            this: NavigationBindings,
             button: HTMLDOMElement
         ): void {
             const navigation = this,
@@ -2329,41 +2010,10 @@ const stockToolsBindings: Record<string, NavigationBindingsOptions> = {
     }
 };
 
-setOptions({
-    navigation: {
-        bindings: stockToolsBindings
-    }
-});
-
-const composedClasses: Array<Function> = [];
-
-/**
- * @private
- */
-function compose(
-    NavigationBindingsClass: typeof NavigationBindings
-): void {
-
-    if (composedClasses.indexOf(NavigationBindingsClass) === -1) {
-        composedClasses.push(NavigationBindingsClass);
-
-        NavigationBindingsClass.prototype.utils = {
-            indicatorsWithAxes,
-            getAssignedAxis,
-            isPriceIndicatorEnabled,
-            manageIndicators
-        };
-    }
-}
-
 /* *
  *
  *  Default Export
  *
  * */
-
-const StockToolsBindings = {
-    compose
-};
 
 export default StockToolsBindings;
